Ambient music indication: Update the position regardless of the visibility state.
This change again ensures that the indication moves up a bit to avoid overlapping
with the loading animation. The intended behavior was lost with this change:
https://github.com/PixelDust-Eleven/android_frameworks_base/commit/f54b30441ff76cadc1c46ea9be283a125718ff0a

Note: The re-positioning works if the indication view is updated (e.g. keyguard or
AoD state has chaged, or the media track is changed).

TODO: Allow to move the indication instantly when the charger state has changed.
- Could be solved be adding a listener which notifies us when the charging state
  was changed and the charging animation is shown by the KeyguardIndicationController.

Signed-off-by: spezi77 <spezi7713@gmx.net>
Signed-off-by: Joey Huab <joey@evolution-x.org>
Signed-off-by: Jayant-Deshmukh <jayantdeshmuk008@gmail.com>
Change-Id: Id581fcf7b1786df59f6d2a671f95791db1d7497c